proj/requirements.txt
# Core dependencies
streamlit>=1.28.0
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.2

# Vector database
chromadb>=0.4.15
chroma-hnswlib>=0.7.3

# Audio processing
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1
pyannote.audio>=3.1.1

# Text processing
langchain>=0.0.350
langchain-community>=0.0.10
nltk>=3.8.1
spacy>=3.7.2

# Data handling
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Web and API
requests>=2.31.0
aiohttp>=3.9.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Environment and config
python-dotenv>=1.0.0
pydantic>=2.5.0

# Utilities
tqdm>=4.66.0
click>=8.1.0
rich>=13.7.0
typer>=0.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Jupyter for evaluation
jupyter>=1.0.0
ipykernel>=6.25.0

# Optional: OpenAI for enhanced responses
openai>=1.3.0

# Optional: HuggingFace for additional models
huggingface-hub>=0.19.0

# Development tools
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0 